requirements.txt
fastapi~=0.63.0
uvicorn~=0.13.4
tortoise-orm[aiomysql]~=0.16.21
python-jose[cryptography]~=3.2.0
python-multipart~=0.0.5
PyYAML~=5.4.1
passlib~=1.7.4